Relocate constants declaration to common dir
